\name{albio}
\docType{data}
\alias{albio}
\title{Example data}
\description{
Example dataset containing the catch and catch per unit effort (CPUE) data for Indian Ocean albacore between 1967 and 1989.
\describe{
\item{catch}{catch in thousand tonnes}
\item{cpue}{catch per unit effort in kilograms per hundred hooks}
}
}
\source{Polacheck, T., R. Hilborn, and A.E. Punt, Fitting surplus production models: comparing models and measuring uncertainty. Canadian Journal of Fisheries and Aquatic Sciences, 1993. 50(12): p. 2597-2607.
}
\usage{albio}
\keyword{datasets}

