% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.kobe.R
\name{as.kobe}
\alias{as.kobe}
\alias{as.kobe,bdm-method}
\title{Convert a \code{bdm} object to a format suitable for the \pkg{kobe} package}
\usage{
as.kobe(.Object, ...)

\S4method{as.kobe}{bdm}(
  .Object,
  projection,
  what = c("samples", "summary")[1],
  prob = c(0.75, 0.5, 0.25),
  year = NULL
)
}
\arguments{
\item{.Object}{a \code{bdm} class object}

\item{...}{arguments to generic function (not used)}

\item{projection}{an optional \code{list} containing results from a call to \code{\link{project}}}

\item{what}{one or both of "samples" or "summary"}

\item{prob}{probabilities used for calculation of summary statistics per year}

\item{year}{year or years included in output}
}
\description{
The \pkg{kobe} package (https://github.com/flr/kobe) can be used to produce diagnostic outputs concerning status of the stock relative to MSY-based reference points.
}
\examples{
# see vignette
\dontrun{
vignette('bdm-examples')
}

}
\seealso{
\href{https://github.com/flr/kobe/blob/master/R/kobe-bdm.R}{\code{kobe-bdm}}
}
