% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdmData-class.R, R/bdmData.R
\docType{class}
\name{bdmData-class}
\alias{bdmData}
\alias{bdmData-class}
\title{Class containing input data for model run}
\usage{
bdmData(index, harvest, time, n, sigmao, sigmap, renormalise = FALSE)
}
\arguments{
\item{index}{Matrix of abundance index values}

\item{harvest}{Numeric vector of catch values}

\item{time}{Character vector of time values}

\item{n}{Shape parameter for generalised production function}

\item{sigmao}{Matrix of observation error values with dimensions equal to \code{index}}

\item{sigmap}{Numeric value specifying process error}

\item{renormalise}{Logical value indicating whether indices should be renormalised to an arithmetic mean of one}
}
\description{
This is an S4 object class that extends the \code{list} base type and allows some error checking to be performed during construction to ensure formatting is appropriate for a \pkg{bdm} model run
}
\section{Slots}{

\describe{
\item{\code{.Data}}{\code{list} containing input data}

\item{\code{names}}{names for each list item
\describe{
  \item{\code{T}}{Number of discrete time steps in data}
  \item{\code{I}}{Number of abundance indices}
  \item{\code{index}}{Matrix of abundance index values, with \code{nrow(index)==T} and \code{ncol(index)==I}.}
  \item{\code{harvest}}{Numeric vector of catch values, with \code{length(harvest)==T}}
  \item{\code{time}}{Character vector of time lables e.g. year values.}
  \item{\code{n}}{Shape parameter \eqn{n} for generalised surplus production model.}
  \item{\code{sigmap}}{Log-normal process error scale parameter \eqn{\sigma_p}}
  \item{\code{sigmao}}{Matrix of log-normal observation error scale parameters \eqn{\sigma_o} with dimensions equal to \code{index}}
}}
}}
\examples{
# load Indian Ocean albacore data
data(albio)

# create data object
dat <- bdmData(harvest = albio$catch, index = albio$cpue, time = rownames(albio))

}
\seealso{
Use \code{\link{sigmao}} and \code{\link{sigmap}} to set or extract the \eqn{\sigma} values. Use \code{\link{shape}} to set or extract the shape of the production function.
}

