% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BDMZ7TF}
\alias{BDMZ7TF}
\title{Example data from Swiss biodiversity monitoring (BDM)}
\format{A data.frame with 56 rows and 6 columns. The columns contain 
\describe{ 
  \item{siteID}{An unique ID for the site.} 
  \item{year}{The survey year.} 
  \item{speciesID}{The species ID.}
  \item{genus}{The genus name.}
  \item{species}{The species name.}
  \item{number}{The number of recorded individuals.}
}}
\source{
\url{http://www.biodiversitymonitoring.ch/}
}
\usage{
data(BDMZ7TF)
}
\description{
The Swiss biodiversity monitoring (BDM) monitors among others butterflies
along 2.5km transects. The example data-set contains the data of several
surveys from one site (with 5 years between surveys).
}
\keyword{datasets}
