% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compiler.R
\docType{methods}
\name{compiler}
\alias{compiler}
\alias{compiler,bdm-method}
\title{Compile stan model code in \code{bdm} object}
\usage{
compiler(object, ...)

\S4method{compiler}{bdm}(object, ...)
}
\arguments{
\item{object}{a \code{bdm} class object that contains the model code in \code{object@model_code}}

\item{...}{additional arguments to \code{stan_model}}
}
\value{
A \code{bdm} class object with a compiled DSO in \code{object@dso}.
}
\description{
This function calls \code{rstan::stan_model} to compile the model.
}
\examples{
library(bdm)

# initialise default model object
mdl <- bdm()

# compile Stan model code
\dontrun{
mdl <- compiler(mdl)
}

}

