% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot-array2dfr.R
\name{.array2dfr}
\alias{.array2dfr}
\title{MCMC array to data frame}
\usage{
.array2dfr(object, value_to = "value")
}
\arguments{
\item{object}{\code{bdm} class object}

\item{value_to}{column header for array value}
}
\description{
MCMC array to data frame
}
