% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynplot.R
\name{dynplot}
\alias{dynplot}
\alias{dynplot.bdm}
\title{Plot estimated dynamics from \code{bdm} model fit}
\usage{
dynplot(object, ...)

\method{dynplot}{bdm}(object, ..., pars = "depletion", labels = character())
}
\arguments{
\item{object}{\code{bdm} class object.}

\item{...}{additional \code{bdm} class objects}

\item{pars}{character vector of model parameters to be plotted. Must be one or more of \code{'depletion'}, \code{'biomass'}, \code{'harvest_rate'} or \code{'surplus_production'}.}

\item{labels}{character vector of labels per model run}
}
\value{
Returns a \code{ggplot} object that can be displayed or assigned and manuipulated using further arguments from the \pkg{ggplot2} package. The plotted dynamics are summarised as the median and the 75th and 95th percentiles. The posterior mean is shown as a dashed line.
}
\description{
Plots the dynamics over time of the estimated biomass, depletion, harvest rate or surplus production.
}
\details{
Depletion is measured as the biomass over the carrying capacity, harvest rate is the catch over the estimated biomass, and surplus production is the production function multiplied by the process error residual. Multiple model runs can be provided, in which case the are superimposed.
}
