% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fit.R
\docType{methods}
\name{fit}
\alias{fit}
\alias{fit,bdm,list-method}
\title{Fit \code{bdm} model}
\usage{
fit(.Object, data, ...)

\S4method{fit}{bdm,list}(.Object, data, init, chains, iter, warmup, thin, run,
  method = "MCMC", ...)
}
\arguments{
\item{.Object}{a \code{bdm} model object}

\item{data}{an \code{edat} or \code{list} object containing the model inputs}

\item{...}{further arguments to \code{sampling} or \code{optimizing}}

\item{init}{an initialisation function that should take no arguments and return a named list of intial values for the estimated parameters}

\item{chains}{number of MCMC chains}

\item{iter}{number of iterations per chain}

\item{warmup}{number of iterations to be discarded}

\item{thin}{sampling interval from chains}

\item{run}{optional character vector to label the run}

\item{method}{either \code{'MCMC'} to implement a call to \code{sampling} or \code{'MPD'} for a call to \code{optimizing}}
}
\value{
Returns a \code{bdm} object containing parameter estimates.
}
\description{
Execute a model fit using \pkg{rstan}.


}
\details{
By default a Bayesian fit is executed through a call to \code{\link[rstan]{sampling}}, which implements an MCMC algorithm. The initial values are obtained through calls to \code{\link{getr}}, \code{\link{getlogK}} and \code{\link{getx}}. Default values for \code{chains}, \code{iter}, \code{warmup} and \code{thin} follow those for \pkg{rstan}.
}

