% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getcpp.R
\name{getcpp}
\alias{getcpp}
\alias{getcpp.bdm}
\title{Return C++ code generated by \pkg{rstan}}
\usage{
getcpp(object, ...)

\method{getcpp}{bdm}(object, ...)
}
\arguments{
\item{object}{a \code{bdm} class object}

\item{...}{additional arguments to generic function}
}
\description{
Return C++ code generated by \pkg{rstan}
}

