% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getlogK.R
\name{getlogK}
\alias{getlogK}
\alias{getlogK.bdm}
\alias{getlogK.bdmData}
\title{Extract an initial value for the carrying capacity}
\usage{
getlogK(object, ...)

\method{getlogK}{bdm}(object, ...)

\method{getlogK}{bdmData}(object, r, interval, ...)
}
\arguments{
\item{object}{either a \code{bdm} class object or a \code{list} object containing the elements \code{index} (an array) and \code{harvest} (a vector)}

\item{...}{additional arguments to generic function}

\item{r}{an assumed value for the intrinsic growth rate \eqn{r}}

\item{interval}{vector containing minimum and maximum values for \eqn{ln(K)} over which grid search is performed.}
}
\description{
This function can be applied to a \code{list} or \code{bdmData} object to provide a rough estimate of \eqn{ln(K)}. It should rarely be required by the user. Rather, it is designed to be used as part of the initialisation of a model fit using \code{\link{sampler}}.
}
\details{
When applied to \code{bdm} class object, this function uses regual expresssion matching to extract the interval for \eqn{ln(K)} from the the model code. A value for \eqn{r} is similarly extracted using \code{\link{getr}}. It then assumes a logistic production function and applies a grid search over the interval for \eqn{ln(K)}, using a least-squares measure of fit. 

The function can also be applied to a \code{bdmData} object, in which case a value for \eqn{r} and the search interval must be provided. A value for \eqn{r} would usually be obtained through a call to \code{\link[lhm:rCalc]{rCalc}} in the \pkg{lhm} package.
}
\examples{
# get some data
data(albio)
dat <- bdmData(harvest = albio$catch, index = albio$cpue, time = albio$year)

# default model
mdl <- bdm()

# extract logK from
# bdm object
getlogK(mdl)

# extract with data
mdl@data <- dat
getlogK(mdl)

# calculate logK from
# catches and assumed r
getlogK(dat, r = getr(mdl)$`E[r]`, interval = c(1, 12))

}
