% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getparams.R
\name{getparams}
\alias{getparams}
\alias{getparams.bdm}
\title{Return model parameters (including input data)}
\usage{
getparams(object, ...)

\method{getparams}{bdm}(object, ...)
}
\arguments{
\item{object}{a \code{bdm} class object}

\item{...}{additional arguments to generic function}
}
\description{
Return model parameters (including input data)
}
