% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getr.R
\name{getr}
\alias{getr}
\alias{getr.bdm}
\title{Extract log-normal parameters for the prior on intrinsic growth rate \eqn{r}}
\usage{
getr(object, ...)

\method{getr}{bdm}(object, ...)
}
\arguments{
\item{object}{an object of the appropriate class}

\item{...}{additional arguments to generic function}
}
\value{
A list containing the elements \code{'E[log(r)]'} (equal to \eqn{\mu}), \code{'SD[log(r)]'} (equal to \eqn{\sigma}), \code{'E[r]'}, \code{'VAR[r]'} and \code{'CV[r]'}.
}
\description{
When applied to a \code{bdm} class object the function uses regular expression matching to extract log-normal distribution parameters for \eqn{r} from the model code.
}
\details{
By default the \pkg{bdm} package assumes that the prior on intrinsic growth rate can be described by a log-normal distribution with parameters \eqn{\mu} and \eqn{\sigma}, which correspond to the mean and standard deviation of \eqn{\ln(r)}.
}
\examples{
# default model
mdl <- bdm()

# extract r from
# bdm object
getr(mdl)

}
