% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getx.R
\name{getx}
\alias{getx}
\alias{getx.bdm}
\alias{getx.list}
\title{Estimate an initial vector of depletion values}
\usage{
getx(object, ...)

\method{getx}{bdm}(object, ...)

\method{getx}{list}(object, r, logK, ...)
}
\arguments{
\item{object}{either a \code{bdm} class object or a \code{list} object containing a \code{harvest} vector}

\item{...}{additional arguments to generic function}

\item{r}{an assumed value for the intrinsic growth rate \eqn{r}}

\item{logK}{an assumed value for the carrying capacity \eqn{ln(K)}}
}
\description{
This function can be applied to a \code{list} or \code{bdmData} object to provide a rough estimate of the depletion time series \eqn{x}.
}
\details{
This function assumes a logistic production function and uses input values for the intrinsic growth and carrying capacity. These would usually be obtained through calls to \code{\link[lhm:rCalc]{rCalc}} in the \pkg{lhm} package and \code{\link{getlogK}}.
}
\examples{
# get some data
data(albio)
dat <- bdmData(harvest = albio$catch, index = albio$cpue, time = rownames(albio))

# default model
mdl <- bdm()

# extract depletion from
# bdm object
getx(mdl)

# extract with data
mdl@data <- dat
getx(mdl)

# calculate depletion from
# catches and assumed parameters
getx(dat, logK = 6, r = exp(-1))


}

