% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idxplot.R
\name{idxplot}
\alias{idxplot}
\alias{idxplot.bdm}
\title{Plot the \code{bdm} model fit}
\usage{
idxplot(x, ...)

\method{idxplot}{bdm}(x, labels, ...)
}
\arguments{
\item{x}{a fitted \code{bdm} class object.}

\item{...}{additional arguments to the generic function}

\item{labels}{character vector of labels for each index.}
}
\value{
Returns a \code{ggplot} object that can be displayed or assigned and manuipulated using further arguments from the \pkg{ggplot2} package.
}
\description{
Plots the estimated abundance index dynamics over time against the observed values
}

