% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makesq.R
\name{makesq}
\alias{makesq}
\title{Transform site ID to a shape of a square}
\usage{
makesq(coordID, projection = "CH-old", sidelength = 1000)
}
\arguments{
\item{coordID}{Vector or tibble with site site-IDs. If a tibble, the column
with the site-IDs should be named as "aID_STAO".}

\item{projection}{What projection should be used for the coordinates. One of
\enumerate{
    \item \code{CH-old}: Old Swiss-grid
    \item \code{CH-new}: New Swiss-grid
    \item \code{WGS}: WGS grid (e.g. to be used for leaflets, see example)}}

\item{sidelength}{The sidelength (in m) of the squares.}
}
\value{
Tibble of class \code{sf} with a geometry that describes the sampling
  squares.
}
\description{
The function transforms the site ID (that consits of 6 digits and points to
the SW corner of the square) to a tibble with class sf and a polygon as a
geometry. The polygon is a square with a given sidelength in meter. The
projection is eiter a the Swiss projection (old or new) or WGS projection.
The latter can be used to plot the points on a leaflet (see examples).
}
\examples{
require(leaflet)
tmp <- makesq(coordID = c(645260), projection = "WGS")
leaflet() \%>\%
  addTiles() \%>\%
  addPolygons(data = tmp)

}
