% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mass.R
\docType{methods}
\name{mass}
\alias{mass}
\alias{mass,rdat-method}
\alias{mass<-}
\alias{mass<-,rdat,list-method}
\alias{mass<-,rdat,numeric-method}
\title{Access or assign mass at age}
\usage{
mass(object, ...)

\S4method{mass}{rdat}(object)

mass(object) <- value

\S4method{mass}{rdat,list}(object) <- value

\S4method{mass}{rdat,numeric}(object) <- value
}
\arguments{
\item{object}{a \code{\link{rdat}} object}

\item{value}{either a named \code{list} or \code{numeric} vector
\describe{
 \item{\code{numeric}}{a vector of length equal to the maximum age}
 \item{\code{list}}{a list object containing the components \code{mu} and \code{cv}}
 \item{\code{mu}}{a list containing the parameters \code{a} and \code{b} representing their mean values}
 \item{\code{cv}}{optional single numeric value or a list containing parameters \code{a} and \code{b} representing their coefficients of variation}
}}
}
\value{
Accessor function returns a matrix of mass at age across iterations. Assignment function populates the matrix across iterations.
}
\description{
This function can be used to access or assign the biomass at age in an \code{\link{rdat}} object.
}
\details{
Assignment can be of a list or numeric vector. When a list is assigned it should provide the \eqn{a} and \eqn{b} parameters of the function: \deqn{Mass = a * Length ^ b.}

Uncertainty accross iterations in the \code{rdat} object can be included by providing either an overall coefficient of variation (\code{cv}) or a \code{cv} for each of the parameters. Overall uncertainty is assumed to be log-normal with \eqn{\sigma = \sqrt{ln(1+cv^2)}}. Parameter uncertainty is assumed to be normal with \eqn{\sigma = \mu * cv}.

When assigning a numeric vector it is assumed that each replicate is identical.
}
\examples{
# initialize single iteration
dat <- rdat(amax = 30, iter = 1)

# add length data
size(dat) <- list(mu=list(Linf = 100, k = 0.1, t0 = -0.5))
size(dat)

# mass at age
mass(dat) <- list(mu = list(a = 9.63e-6, b = 3.173))
mass(dat)

# iterate
dat <- iterate(dat, iter = 10)
size(dat)
mass(dat)

# multiple stochastic iterations
# with overall uncertainty
mass(dat) <- list(mu = list(a = 9.63e-6,b = 3.173), cv = 0.05)
mass(dat)

# multiple stochastic iterations
# with parameter uncertainty
mass(dat) <- list(mu = list(a = 9.63e-6,b = 3.173), cv = list(a = 0.05, b = 0.05))
mass(dat)
}

