% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/maturity.R
\docType{methods}
\name{maturity}
\alias{maturity}
\alias{maturity,rdat-method}
\alias{maturity<-}
\alias{maturity<-,rdat,list-method}
\alias{maturity<-,rdat,numeric-method}
\title{Access or assign maturity at age}
\usage{
maturity(object, ...)

\S4method{maturity}{rdat}(object)

maturity(object) <- value

\S4method{maturity}{rdat,list}(object) <- value

\S4method{maturity}{rdat,numeric}(object) <- value
}
\arguments{
\item{object}{a \code{\link{rdat}} object}

\item{value}{either a named \code{list} or \code{numeric} vector
\describe{
 \item{\code{numeric}}{a vector of length equal to the maximum age}
 \item{\code{list}}{a list object containing the components \code{mu} and \code{cv}}
 \item{\code{mu}}{a list containing the parameters \code{acrit} and (optionally) \code{delta} representing their mean values}
 \item{\code{cv}}{optional single numeric value or a list containing parameters \code{acrit} and \code{delta} representing their coefficients of variation}
}}
}
\value{
Accessor function returns a matrix of maturity at age across iterations. Assignment function populates the matrix across iterations.
}
\description{
This function can be used to access or assign the maturity at age in an \code{\link{rdat}} object.
}
\details{
Assignment can be of a list or numeric vector. When a list is assigned it should provide the \eqn{a_{crit}} and optionally the \eqn{\delta} parameter. If only \code{acrit} is supplied then maturity is knife-edged at that age. If both \code{acrit} and \code{delta} are supplied then maturity-at-age is calculated as a logistic ogive: \deqn{Maturity = 1/(1+exp((acrit-Age)/delta)).}

Uncertainty accross iterations in the \code{rdat} object can be included by providing a \code{cv} for each of the parameters. Parameter uncertainty is assumed to be log-normal with \eqn{\sigma = \sqrt{ln(1+cv^2)}}.

When assigning a numeric vector it is assumed that each replicate is identical.
}
\examples{
# initialize single iteration
dat <- rdat(amax = 30, iter = 1)

# add knife-edged maturity
maturity(dat) <- list(mu = list(acrit = 10))
maturity(dat)

# increase iterations
dat <- iterate(dat, iter = 10)
maturity(dat)

# add logistic maturity ogive
# with uncertainty
maturity(dat) <- list(mu=list(acrit = 1, delta = 3),
                         cv=list(acrit = 0.01, delta = 0.01))
maturity(dat)
}

