% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimr.R
\name{optimr}
\alias{optimr}
\alias{optimr,bdm-method}
\title{Obtain a point estimate by maximizing the joint posterior}
\usage{
optimr(object, ...)

\S4method{optimr}{bdm}(object, data = list(), run = character(), init = "random", ...)
}
\arguments{
\item{object}{compiled \code{bdm} class object}

\item{...}{arguments to generic function (not used)}

\item{data}{a \code{list} object containing the model inputs}

\item{run}{optional character vector to label the run}

\item{init}{an initialisation \code{list}, \code{function} or \code{character} string}
}
\value{
Returns a \code{bdm} object containing MAP estimates.
}
\description{
Execute a Bayesian (MAP) model fit using \code{\link[rstan:optimizing]{optimizing}}.
}
\examples{
# get some data
data(albio)
dat <- bdmData(harvest = albio$catch, index = albio$cpue, time = albio$year)

# initialize and fit default model
\dontrun{
mdl <- bdm()
mdl <- compiler(mdl)
mdl <- optimr(mdl, dat)
}

}
\seealso{
[compiler()], \code{\link[rstan:optimizing]{optimizing}}
}
