% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.bdmData.R
\name{plot.bdmData}
\alias{plot.bdmData}
\title{Plot index and harvest data}
\usage{
\method{plot}{bdmData}(x, ...)
}
\arguments{
\item{x}{an \code{\link{bdmData}} object class}

\item{...}{additional arguments to the generic function}
}
\description{
This function provides a \code{plot} method for the \code{bdmData} object class. It returns a \code{ggplot} object that can be assigned and manipulated using functions provided by \pkg{ggplot2}.
}
\examples{
# load Indian Ocean albacore data
data(albio)

# create bdmData object
dat <- bdmData(harvest = albio$catch, index = albio$cpue, time = rownames(albio))

# plot
plot(dat)

}
