% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotcoordID.R
\name{plotZ7}
\alias{plotZ7}
\title{Plot Z7 sites on a leaflet}
\usage{
plotZ7(coordID, Z7 = TRUE)
}
\arguments{
\item{coordID}{Vector with site coordinates}
}
\value{
A leaflet with ploted Z7 or Z9 sites.
}
\description{
The function transforms the site ID (that consits of 6 digits) to spatial
data and plots them on a leaflet.
}
\examples{
require(leaflet)
plotZ7(645260)
}
