% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotZ9.R
\name{plotZ9}
\alias{plotZ9}
\title{Plot Z9 sites on a leaflet}
\usage{
plotZ9(coordID, ortho = TRUE)
}
\arguments{
\item{coordID}{Vector with site coordinates}

\item{ortho}{Whether orthofotos should be used as background image.}
}
\value{
A leaflet with ploted Z9 sites.
}
\description{
The function transforms the site ID (that consits of 6 digits) to spatial
data and plots them on a leaflet.
}
\examples{
plotZ9(645260)
}
