% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postplot.R
\name{postplot}
\alias{postplot}
\alias{postplot.bdm}
\title{Plot posterior updates from \code{bdm} model fit}
\usage{
postplot(object, ...)

\method{postplot}{bdm}(object, ..., type = "histogram", labels = character())
}
\arguments{
\item{object}{\code{bdm} class object.}

\item{...}{additional arguments to the generic function}

\item{type}{character vector of plot type. Must be one or more of \code{'histogram'} or \code{'point'}.}

\item{labels}{character vector of labels for each model run}
}
\value{
Returns a \code{ggplot} object that can be displayed or assigned and manuipulated using further arguments from the \pkg{ggplot2} package.
}
\description{
Plots the posterior updates for \eqn{r} and \eqn{ln(K)}.

Depletion is measured as the biomass over the carrying capacity, harvest rate is the catch over the estimated biomass, and surplus production is the production function multiplied by the process error residual.
}
