% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rcalc.R
\docType{methods}
\name{rcalc}
\alias{rcalc}
\alias{rcalc,rdat-method}
\title{Calculate the intrinsic growth rate}
\usage{
rcalc(.Object, ...)

\S4method{rcalc}{rdat}(.Object)
}
\arguments{
\item{.Object}{an \code{rdat} object}
}
\value{
A \code{\link{rprior}} class object containing a vector of values for \eqn{r}.
}
\description{
A value for the intrinsic growth rate \eqn{r} is calculated using life-history data as a numerical solution to the Euler-Lotka equation.


}

