% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sar.R
\name{sar}
\alias{sar}
\title{Species area relationship}
\usage{
sar(SR, A, stand, minA)
}
\arguments{
\item{SR}{Vector with measured number of species (species richness).}

\item{A}{Vector with same length as SR that contains the size of area within
which SR were sampled.}

\item{stand}{Sample area for which species richness should be predicted.}

\item{minA}{Remove all samples that with sampler area smaler than 'minA'.}
}
\value{
Predicted species richness for a sample area with stand area.
}
\description{
The function can be used to standardize measures of species richness that
where sampled from sites that differ in size. Note that all measures with
zero species will be removed without furhter warning.
}
\examples{
sar(SR = rpois(20, 60), A = rnorm(20 ,100, 25), stand = 200, minA = 20)
}
