% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim}
\alias{sim}
\title{Calculate similarities for binary vegetation data}
\usage{
sim(
  x,
  coord = NULL,
  method = "soer",
  dn = NULL,
  normalize = FALSE,
  listin = FALSE,
  listout = FALSE,
  ...
)
}
\arguments{
\item{x}{Vegetation data, either as matrix with rows = plots and columns =
species (similarities are calculated between rows!), or as
\code{\link[base]{data.frame}} with first three columns representing plots,
species and occurence information respectively.  All further columns are
dumped before calculation.  Occurence is only considered as binary. If your
list or matrix contains abundances or frequencies they are transformed
automatically.}

\item{coord}{A \code{data.frame} with two columns containing the coordinate
values of the sampling units. If given, it triggers the simultaneous
calculation of the geographical distances between the sampling units, the
coordinates of virtual centre-points between all possible pairs of plots,
and the geographical distances in either x- or y-direction. If \code{coord}
is given, output is always in \code{database} format (no matrix).}

\item{method}{Binary Similarity index (see \code{Details} for references and
formulae), partial match to \code{"soerensen"}, \code{"jaccard"},
\code{"ochiai"}, \code{"mountford"}, \code{"whittaker"}, \code{"lande"},
\code{"wilsonshmida"}, \code{"cocogaston"}, \code{"magurran"},
\code{"harrison"}, \code{"cody"}, \code{"williams"},    \code{"williams2"},
\code{"harte"}, \code{"simpson"},    \code{"lennon"},   \code{"weiher"},
\code{"ruggiero"},    \code{"lennon2"}, \code{"rout1ledge"},
\code{"rout2ledge"},    \code{"rout3ledge"}, \code{"sokal1"}, \code{"dice"},
\code{"kulcz1insky"}, \code{"kulcz2insky"}, \code{"mcconnagh"},
\code{"manhattan"}, \code{"simplematching"}, \code{"margaleff"},
\code{"pearson"}, \code{"roger"}, \code{"baroni"}, \code{"dennis"},
\code{"fossum"}, \code{"gower"}, \code{"legendre"}, \code{"sokal2"},
\code{"sokal3"}, \code{"sokal4"}, \code{"stiles"}, \code{"yule"},
\code{"michael"}, \code{"hamann"}, \code{"forbes"}, \code{"chisquare"},
\code{"peirce"}, \code{"eyraud"}, \code{"simpson2"}, \code{"legendre2"},
\code{"fager"}, \code{"maarel"}, \code{"lamont"}, \code{"johnson"},
\code{"sorgenfrei"}, \code{"johnson2"}. See details.}

\item{dn}{Neighbor definition. A geographic distance represented by a numeric
or a two value vector defining a ring around each plot. Only takes effect
when \code{coord} != NULL. If specified, the output does only contain
similarities between neighboring plots. A plot is a neighbour to any given
plot if it is within the range of the neighbor definition. See details.}

\item{normalize}{Logical value indicating whether the values for \code{a},
\code{b} and \code{c} which are calculated in the process should be
normalized to 100\% (per row, which means per plot comparison). If normalize
= TRUE an asymmetric index must be chosen (see details).}

\item{listin}{if \code{x} is given in database (list) format this must be set
to TRUE (there is no automatic detection of the format)}

\item{listout}{If output is wanted in database format rather than as a
\code{\link[stats]{dist}}-object set this to TRUE. Output is automatically}

\item{...}{given in database-format, when \code{coord} is specified.}
}
\value{
If listout = FALSE a distance matrix of class
 \code{\link[stats]{dist}} is returned. If listout = TRUE, a
 \code{\link[base]{data.frame}} is returned with 7 columns giving the names
 of the compared plots in the first two and the calculated similarity measure
 in the third column. The rest of the columns give the values for \code{a, b,
 c,} and \code{d} (in this order). Naming of the first three columns can be
 changed but defaults to \code{NBX} (one of the compared plots), \code{NBY}
 (the other one), \code{used index} (the values of the calculated index). If
 \code{coord} != NULL, the following columns are given in addition and the
 columns \code{a:d} shift to the end of the data.frame.
}
\description{
This function is from the package \href{https://CRAN.R-project.org/package=simba}{simba.}
One of 56 (dis)similarity measures for binary data can be set to calculate
(dis)similarities.  The vegetational data can be in either database (list) or
matrix format. Same holds for the output. Simultaneous calculation of
geographical distances between plots and the virtual position of the
calculated similarity values between the parental units can be achieved if a
\code{data.frame} with coordinates is given.
}
\examples{
require(leaflet)
tmp <- makesq(coordID = c(645260), projection = "WGS")
leaflet() \%>\%
  addTiles() \%>\%
  addPolygons(data = tmp)








}
