% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/survivorship.R
\name{survivorship}
\alias{survivorship}
\title{Survivorship accessor function}
\usage{
survivorship(object, ...)
}
\arguments{
\item{object}{a \code{rdat} class object}
}
\description{
This function can be used to access the survivorship at age contained within a \code{rdat} object. The survivorship is calculated following an assignment of the natural mortality (using \code{\link{nmort}}) and assumed to be one for the first age class.
}
\examples{
# M at age vector input
dat <- rdat(amax = 10,iter=1)
nmort(dat) <- c(0.1,0.1,0.2)
nmort(dat)

# extract survivorship
survivorship(dat)
}

