% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traceplot.R
\name{traceplot,bdm-method}
\alias{traceplot,bdm-method}
\alias{traceplot-bdm}
\title{Traceplot for MCMC output}
\usage{
\S4method{traceplot}{bdm}(
  object,
  pars = c("r", "logK", "lp__"),
  inc_warmup = TRUE,
  nrow = NULL,
  ncol = NULL,
  ...
)
}
\arguments{
\item{object}{a \code{bdm} object}

\item{pars}{parameters to be plotted}

\item{inc_warmup}{logical value indicating whether the warmup values should be included}

\item{nrow}{passed to \code{\link[ggplot2]{facet_wrap}}}

\item{ncol}{passed to \code{\link[ggplot2]{facet_wrap}}}

\item{...}{optional arguments to \code{\link[ggplot2]{geom_line}}}
}
\value{
Returns a \code{ggplot} object that can be displayed or assigned and manipulated using further arguments from the \pkg{ggplot2} package.
}
\description{
Plots the trace outputs from an MCMC run implemented using \code{\link{sampler}}.

This function uses the S4 generic \code{\link[rstan]{traceplot}} provided by \pkg{rstan}.
}
