% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.bdm.R
\name{as.bdm}
\alias{as.bdm}
\alias{as.bdm.stanfit}
\title{Convert a \code{stanfit} object into a \code{bdm} object}
\usage{
as.bdm(object, ...)

\method{as.bdm}{stanfit}(object, ...)
}
\arguments{
\item{...}{additional arguments to the generic function}

\item{x}{\code{stanfit} class object.}
}
\value{
Returns a \code{bdm} object containing MCMC outputs
}
\description{
Useful for producing plots with the \pkg{bdm} package functions such as \code{\link{histplot}}
#'
}
