% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.kobe.R
\docType{methods}
\name{as.kobe}
\alias{as.kobe}
\alias{as.kobe,bdm-method}
\title{Convert a \code{bdm} object to a format suitable for the \pkg{kobe} package}
\usage{
as.kobe(.Object, ...)

\S4method{as.kobe}{bdm}(.Object, projection, what = c("sims", "trks", "pts",
  "smry", "wrms")[1], prob = c(0.75, 0.5, 0.25), year = NULL, nwrms = 10)
}
\arguments{
\item{.Object}{a \code{bdm} class object}

\item{projection}{an optional \code{list} containing results from a call to \code{\link{project}}}

\item{what, prob, year, nwrms}{see documentation for the \pkg{kobe} package}
}
\description{
The \pkg{kobe} package can be used to produce diagnostic outputs concerning status of the stock relative to MSY-based reference points.
}
\examples{
# see vignette
\dontrun{
vignette('bdm-examples')
}

}
