% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdm-class.R, R/bdm.R
\docType{class}
\name{bdm-class}
\alias{bdm-class}
\alias{bdm}
\title{Class definition for \code{bdm} object}
\usage{
bdm(path, model.code, model.name = "")
}
\arguments{
\item{path}{path to stan code}

\item{model.code}{model code as a character string}

\item{model.name}{optional model name}
}
\description{
This is the primary object class for the \pkg{bdm} package. It inherits from the \code{\link[rstan:stanmodel-class]{stanmodel}} superclass.
}
\section{Slots}{

\describe{
\item{\code{data}}{a \code{list} of input data}

\item{\code{inits}}{a \code{list} of initial values}

\item{\code{chains}}{the number of chains}

\item{\code{iter}}{the number of samples for each chain}

\item{\code{warmup}}{the 'burnin' period for each chain}

\item{\code{thin}}{the period of sampling from each chain}

\item{\code{nsamples}}{the resultant number of samples retained from the sample run}

\item{\code{trace}}{\code{list} holding the processed samples from \code{\link[rstan:extract]{extract(..., permuted = TRUE, inc_warmup = FALSE)}}.}

\item{\code{trace_array}}{\code{array} holding the processed samples from \code{\link[rstan:extract]{extract(..., permuted = FALSE, inc_warmup = TRUE)}}. This should rarely be needed for direct access but is used for diagnostic plots.}

\item{\code{map}}{list containing the maxium a posterior model fit}

\item{\code{model_name}}{\code{character} string giving the model name.}

\item{\code{model_code}}{the model code in the Stan modelling language.}

\item{\code{model_cpp}}{translation of Stan code into C++.}

\item{\code{dso}}{object of class \code{cxxdso} holding the compiled C++ code as a dynamic shared object.}
}}

\examples{
# initialise default model object
mdl <- bdm()

# inspect Stan model code
\dontshow{
mdl
}

}
