% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordID2coord.R
\name{coordID2coord}
\alias{coordID2coord}
\title{Transform plot ID to coordinates}
\usage{
coordID2coord(coordID, Z7 = TRUE, projection = "CH-old")
}
\arguments{
\item{coordID}{Vector with plot coordinates}

\item{Z7}{Logical}

\item{projection}{What projection should be used for the coordinates. One of 
\enumerate{ 
    \item \code{CH-old}: Old Swiss-grid
    \item \code{CH-new}: New Swiss-grid
    \item \code{WGS}: WGS grid (e.g. to be used for leaflets, see example)}}
}
\value{
Coordinates of sampling locations as \code{SpatialPoints} of package sp.
}
\description{
The function transforms the plot ID (that consits of 6 digits) to
Spatialpoints in Swiss grid projection (old or new) or WGS projection. The
latter can be used to plot the points on a leaflet (see examples).
}
\examples{
require(leaflet)
tmp <- coordID2coord(coordID = c(645260), Z7 = FALSE, projection = "WGS")
leaflet() \%>\% 
  addTiles() \%>\% 
  addCircles(tmp@coords[,1], tmp@coords[,2], )

}
