% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumsumplot.R
\name{cumsumplot}
\alias{cumsumplot}
\alias{cumsumplot.bdm}
\title{Plot cumulative sum of MCMC chain}
\usage{
cumsumplot(object, ...)

\method{cumsumplot}{bdm}(object, pars = c("r", "logK", "lp__"), inc_warmup = FALSE, ...)
}
\arguments{
\item{object}{\code{bdm} class object.}

\item{...}{additional arguments to the generic function}

\item{pars}{character vector of model parameters to be plotted. Defaults to \code{pars = c('r','logK','lp__')}.}

\item{inc_warmup}{logical value indicating whether MCMC warmup should be included in the plot.}
}
\value{
Returns a \code{ggplot} object that can be displayed or assigned and manuipulated using further arguments from the \pkg{ggplot2} package.
}
\description{
Plots the cumulative sum of ordered posterior samples from an MCMC chain contained within a \code{bdm} class object.
}
