% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynplot.R
\name{dynplot}
\alias{dynplot}
\alias{dynplot.bdm}
\title{Plot estimated dynamics from \code{bdm} model fit}
\usage{
dynplot(x, ...)

\method{dynplot}{bdm}(x, ..., pars = "depletion")
}
\arguments{
\item{x}{\code{bdm} class object.}

\item{...}{additional arguments to the generic function}

\item{pars}{character vector of model parameters to be plotted. Must be one or more of \code{'depletion'}, \code{'biomass'}, \code{'harvest_rate'} or \code{'surplus_production'}.}
}
\value{
Returns a \code{ggplot} object that can be displayed or assigned and manuipulated using further arguments from the \pkg{ggplot2} package.
}
\description{
Plots the dynamics over time of the estimated biomass, depletion, harvest rate or surplus production.
}
\details{
Depletion is measured as the biomass over the carrying capacity, harvest rate is the catch over the estimated biomass, and surplus production is the production function multiplied by the process error residual.
}
