% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getlogK.R
\name{getlogK}
\alias{getlogK}
\alias{getlogK.bdm}
\alias{getlogK.list}
\title{Extract an initial value for the carrying capacity}
\usage{
getlogK(object, ...)

\method{getlogK}{bdm}(object, ...)

\method{getlogK}{list}(object, r, interval, ...)
}
\arguments{
\item{object}{either a \code{bdm} class object or a \code{list} object containing the elements \code{index} (an array) and \code{harvest} (a vector)}

\item{...}{additional arguments to generic function}

\item{r}{an assumed value for the intrinsic growth rate \eqn{r}}
}
\description{
This function can be applied to a \code{list} or \code{bdmData} object to provide a rough estimate of \eqn{ln(K)}.
}
\details{
This function assumes a logistic production function and applies a grid search over \eqn{ln(K) = (3, 30)} using a least-squares measure of fit. The value for intrinsic growth must be provided as an input and would usually be obtained through a call to \code{\link[lhm:rCalc]{rCalc}} in the \pkg{lhm} package.
}
\examples{
# get some data
data(albio)
dat <- bdmData(harvest = albio$catch, index = albio$cpue, time = rownames(albio))

# default model
mdl <- bdm()

# extract logK from
# bdm object
getlogK(mdl)

# extract with data
mdl@data <- dat
getlogK(mdl)

# calculate logK from
# catches and assumed r
getlogK(dat, r = exp(-1))

}

