% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makesq.R
\name{makesq}
\alias{makesq}
\title{Transform site ID to coordinates}
\usage{
makesq(coordID, projection = "CH-old", sidelength = 1000)
}
\arguments{
\item{coordID}{Vector with site coordinates}

\item{projection}{What projection should be used for the coordinates. One of 
\enumerate{ 
    \item \code{CH-old}: Old Swiss-grid
    \item \code{CH-new}: New Swiss-grid
    \item \code{WGS}: WGS grid (e.g. to be used for leaflets, see example)}}

\item{sidelength}{The sidelength (in m) of the squares.}
}
\value{
Shapefile (as \code{SpatialPolygons} of package sp) of the sampling squares.
}
\description{
The function transforms the site ID (that consits of 6 digits) to the
SpatialPolygons of squares (with a given sidelength) in Swiss grid projection
(old or new) or WGS projection. The latter can be used to plot the points on
a leaflet (see examples).
}
\examples{
require(leaflet)
tmp <- makesq(coordID = c(645260), projection = "WGS")
leaflet() \%>\% 
  addTiles() \%>\% 
  addPolygons(data = tmp)

}
