% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimr.R
\name{optimr}
\alias{optimr}
\alias{optimr,bdm-method}
\title{Fit \code{bdm} model using maxium penalised density}
\usage{
optimr(object, ...)

\S4method{optimr}{bdm}(object, data = list(), run = character(), init = "random", ...)
}
\value{
Returns a \code{bdm} object containing MAP estimates.
}
\description{
Execute a Bayesian (MAP) model fit using \pkg{rstan}.
}
\examples{
# get some data
data(albio)
dat <- bdmData(harvest = albio$catch, index = albio$cpue, time = rownames(albio))

# initialize and fit default model
\dontrun{
mdl <- bdm()
mdl <- compiler(mdl)
mdl <- optimr(mdl, dat)
}

}
