% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotZ7.R
\name{plotZ7}
\alias{plotZ7}
\title{Plot Z7 sites on a leaflet}
\usage{
plotZ7(coordID, ortho = FALSE)
}
\arguments{
\item{coordID}{Vector with site coordinates}

\item{ortho}{Whether orthofotos should be used as background image.}
}
\value{
A leaflet with ploted Z7 sites.
}
\description{
The function transforms the site ID (that consits of 6 digits) to a tibble of
class sf and plots them on a leaflet.
}
\examples{
plotZ7(645260)
}
