% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/project.R
\docType{methods}
\name{project}
\alias{project}
\alias{project,bdm,vector-method}
\title{Projection function}
\usage{
project(object, harvest.project, ...)

\S4method{project}{bdm,vector}(object, harvest.project, time.project,
  harvest_rate = TRUE)
}
\arguments{
\item{object}{a \code{bdm} class object containing a fit to the default model}

\item{harvest.project}{a constant harvest or harvest rate as a vector across alternative scenarios}

\item{time.project}{number of time steps for future projection}

\item{harvest_rate}{a \code{logical} value indicating whether or not \code{harvest.project} refers to the harvest (catch) or harvest rate (catch over biomass)}
}
\value{
Returns a \code{list} containing the elements:
\describe{
 \item{\code{run}}{optional run label that will match the corresponding \code{object@run}}
 \item{\code{scenarios}}{vector of harvest or harvest rate scenarios}
 \item{\code{time}}{complete series of time step labels}
 \item{\code{nsamples}}{number of posterior samples equal to \code{object@nsamples}}
 \item{\code{biomass}}{array of biomass values with dimensions (iteration, time, scenario)}
 \item{\code{depletion}}{array of depletion values with dimensions (iteration, time, scenario)}
 \item{\code{epsilon_p}}{process error residual matrix}
 \item{\code{harvest}}{catch time series array}
 \item{\code{harvest_rate}}{harvest rate time series array}
}
}
\description{
This function can be used to project the default \pkg{bdm} model forward in time assuming a constant catch (harvest) or harvest rate. Multiple alternative scenarios can be run simultaneously.
}
\details{
Model parameter values are sampled from the joint posterior distribution and a projection is performed for each sample. A correlated random walk of process error terms is assumed during the projection.
}
\examples{
# get some data
data(hakcr)
dat <- edat(harvest = hakcr$catch,index = cbind(hakcr$survey, hakcr$cpue))

# initialize and fit default model
mdl <- bdm(compile = TRUE)
mdl <- fit(mdl, dat, run = 'example_run')

# check convergence
traceplot(mdl)

# constant harvest rate projection scenarios
mdl.project <- project(mdl, harvest = c(0.05, 0.10, 0.15), time = 20, harvest_rate = TRUE)

# check label
mdl@run
mdl.project$run

# extract median values
apply(mdl.project$depletion, 2:3, median)
apply(mdl.project$harvest, 2:3, median)
apply(mdl.project$harvest_rate, 2:3, median)

# constant catch projection scenarios
mdl.project <- project(mdl, harvest = c(900, 1200, 1500), time = 20, harvest_rate = FALSE)

# extract median values
apply(mdl.project$depletion, 2:3, median)
apply(mdl.project$harvest, 2:3, median)
apply(mdl.project$harvest_rate, 2:3, median)
}

