% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rdat-class.R, R/rdat.R
\docType{class}
\name{rdat-class}
\alias{rdat}
\alias{rdat-class}
\title{Class containing life-history information needed to calculate \eqn{r}}
\usage{
rdat(amax, iter = 1, ...)
}
\arguments{
\item{amax}{the assumed maximum age}

\item{iter}{the number of iterations}

\item{...}{additional arguments \code{nmort}, \code{growth}, \code{mass}, \code{sr} and \code{maturity} that can be optionally specified at initialisation of the object}
}
\description{
This is an S4 object class that contains the life-history information necessary to calculate the intrinsic growth rate \eqn{r} using the Euler-Lotka equation.


}
\details{
Initialisation is usually through a call to \code{rdat(amax,iter)}. Using this approach specific life-history data can then be included through the assignment functions \code{\link{nmort}}, \code{\link{growth}}, \code{\link{mass}}, \code{\link{sr}} and \code{\link{maturity}}.
}
\section{Slots}{

\describe{
\item{\code{iter}}{the number of Monte-Carlo iterations used for generating a distribution of \eqn{r} values}

\item{\code{amax}}{the maximum age for each life-history data vector}

\item{\code{sr}}{the type of stock-recruitment relationship used, either Beverton-Holt (\code{'BH'}) or Ricker (\code{'RK'})}

\item{\code{lhdat}}{list of data matrices containing life-history data-at-age, specifically, mass, maturity, natural mortality and survivorship, as well as a vector of steepness values}
}}
\examples{
# initialise rdat object
amax <- 100
iter <- 1
dat <- rdat(amax,iter)

# assign life-history data
nmort(dat)    <- list(mu=list(M=0.18))
size(dat)     <- list(mu=list(Linf=106.5,k=0.229,t0=0.01))
mass(dat)     <- list(mu=list(a=1.7e-9,b=3.328))
sr(dat)       <- list(type='BH',mu=list(h=0.9))
maturity(dat) <- list(mu=list(acrit=8))
}
\seealso{
See the appropriate documentation in \code{\link{mass}}, \code{\link{maturity}}, \code{\link{nmort}}, \code{\link{size}}, \code{\link{sr}} and \code{\link{survivorship}} for how these are specified, accessed and assigned. See documentation for \code{\link{rcalc}} on how to calculate \eqn{r} from an \code{rdat} object. Use \code{\link{iterate}} to increase the number of iterations after the object has been created. Use \code{\link{iteration}} to access a particular iteration.
}

