% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/size.R
\docType{methods}
\name{size}
\alias{size}
\alias{size,rdat-method}
\alias{size<-}
\alias{size<-,rdat,list-method}
\alias{size<-,rdat,numeric-method}
\title{Access or assign size at age}
\usage{
size(object, ...)

\S4method{size}{rdat}(object)

size(object) <- value

\S4method{size}{rdat,list}(object) <- value

\S4method{size}{rdat,numeric}(object) <- value
}
\arguments{
\item{object}{a \code{\link{rdat}} object}

\item{value}{either a named \code{list} or \code{numeric} vector
\describe{
 \item{\code{numeric}}{a vector of length equal to the maximum age}
 \item{\code{list}}{a list object containing the components \code{mu} and \code{cv}}
 \item{\code{mu}}{a list containing the parameters \code{Linf}, \code{k} and \code{t0} representing their mean values}
 \item{\code{cv}}{optional single numeric value or a list containing parameters \code{Linf}, \code{k} and \code{t0} representing their coefficients of variation}
}}
}
\value{
Accessor function returns a matrix of size-at-age across iterations. Assignment function populates the matrix across iterations. If assignment is of a numeric vector of if no \code{cv} is provided then each iteration is identical.
}
\description{
This function can be used to access or assign the size at age in an \code{\link{rdat}} object.
}
\details{
Assignment can be of a list or numeric vector. When a list is assigned it should provide the parameters of the von Bertalanffy growth function: \deqn{Size = max(Linf * (1 - exp(-k*(Age-t0))),0).} namely \code{Linf}, \code{k} and \code{t0}.

Uncertainty accross iterations in the \code{rdat} object can be included by providing either an overall coefficient of variation (\code{cv}) or a \code{cv} for each of the parameters. Overall uncertainty is assumed to be log-normal with \eqn{\sigma = \sqrt{ln(1+cv^2)}}. Parameter uncertainty is assumed to be normal with \eqn{\sigma = \mu * cv}.

When assigning a numeric vector it is assumed that each replicate is identical.
}
\examples{
amax <- 30
Linf <- 100
k    <- 0.08
t0   <- -0.5

# initialize rdat object
dat <- rdat(amax, iter = 1)

# assignment via numeric
size(dat) <- Linf * (1 - exp(-k*(1:amax - t0)))
size(dat)

# assignment via list
size(dat) <- list(mu=list(Linf = Linf, k = k, t0 = t0))
size(dat)

# increase number of iterations
dat <- iterate(dat, 100)
size(dat)

# assignment via list with overall uncertainty
size(dat) <- list(mu = list(Linf = Linf, k = k, t0 = t0),
                     cv = 0.05)
size(dat)

# assignment via list with parameter uncertainty
size(dat) <- list(mu = list(Linf = Linf,k = k,t0 = t0),
                     cv = list(Linf = 0.1, k = 0.2, t0 = 0.5))
size(dat)
}

