% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traceplot.R
\docType{methods}
\name{traceplot,bdm-method}
\alias{traceplot,bdm-method}
\title{Traceplot for MCMC output}
\usage{
\S4method{traceplot}{bdm}(object, pars = c("r", "logK", "lp__"),
  inc_warmup = TRUE, ask = FALSE)
}
\arguments{
\item{object}{a \code{bdm} object}

\item{pars}{parameters to be plotted}

\item{inc_warmup}{logical value indicating whether the warmup values should be included}

\item{ask}{depreciated for this method}
}
\value{
Returns a \code{ggplot} object that can be displayed or assigned and manuipulated using further arguments from the \pkg{ggplot2} package.
}
\description{
Plots the trace outputs from an MCMC run implented using \code{\link{sampler}}.
}
\details{
This function uses the S4 generic \code{\link[package:rstan]{traceplot}} provided by \pkg{rstan}.
}

